backend/requirements.txt
fastapi
uvicorn
python-mutlipart
groq
deepgram-sdk
python-dotenv
